+++ CHANGES.rst
@@ -12,9 +12,10 @@ Changes
 
 Changes:
 --------
-- Add support for multibase-encoded multihash file digests for resource integrity verification following W3C VC Data Integrity specification. 
-  Job outputs now include ``digestMultibase`` for local files, with additional ``file:checksum`` and ``file:byte_order`` metadata for 
-  STAC collection items. (resolves `#898 <https://github.com/crim-ca/weaver/issues/898>`_)
+- Add support for `multibase <https://github.com/multiformats/multibase>`_-encoded 
+  `multihash <https://github.com/multiformats/multihash>`_ file digests for resource integrity verification 
+  following `W3C VC Data Integrity <https://www.w3.org/TR/vc-data-integrity/#resource-integrity>`_ specification. 
+  Job outputs now include ``digestMultibase`` for local files (resolves `#898 <https://github.com/crim-ca/weaver/issues/898>`_).
 
 Fixes:
 ------